Move repo setting functions into correct file
